requirements.txt
boto3==1.23.2
Flask==2.1.2
python-dotenv==0.20.0
